+++ test/option_popup/stubs/HalGPIO.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class HalGPIO {
+ public:
+  bool hasTouch() const { return false; }
+};
+
+inline HalGPIO gpio;